+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 * Add `GdsApi::PublishingApiV2#lookup_content_id` and `GdsApi::PublishingApiV2#lookup_content_id`
+* Add `publishing_api_has_lookups` test helpers
 
 # 29.5.0